# swagger.yaml
openapi: '3.0.0'
info:
  title: Express Forum APP
  description: simple Express Forum APP API
  version: '1.0.0'
servers:
  - url: /api/v1
paths:
...
  /auth/reset-password/{token}:
    patch:
      tags:
        - auth
      summary: reset your password
      operationId: reset-password
      parameters:
        - name: 'token'
          in: 'path'
          description: 'reset password token'
          required: true
          type: 'string'
      requestBody:
        content:
          application/json:
            schema:
              type: 'object'
              properties:
                password:
                  type: 'string'
                  default: 'mynewpassword'
                passwordConfirmation:
                  type: 'string'
                  default: 'mynewpassword'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          description: Token is invalid or has expired
        '422':
          $ref: '#/components/responses/UnprocessableError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Success:
      description: success
...
    UnprocessableError:
      description: Invalid inputs
...